Update number_gassing_game_.py
This is a fun guessing game. 
First, the user inputs a start and end number. 
Then the computer  generates a random number within this range for the user to guess. 
You have a maximum of 10 attempts to guess the number; if you don't succeed, the computer reveals the number, and you lose the game.